requirements.txt
# Core Dependencies
portia-sdk-python>=0.7.2
hume>=0.9.0
python-dotenv>=1.0.0

# Audio Processing
pyaudio>=0.2.11
numpy>=1.24.0

# Async Support
asyncio-mqtt>=0.13.0
websockets>=11.0.0

# Data Processing
pandas>=2.0.0
dataclasses-json>=0.6.0

# Logging and Monitoring
structlog>=23.0.0
rich>=13.0.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: For enhanced audio processing
scipy>=1.11.0
librosa>=0.10.0

